continuing nostdification
